requirements.txt
bleach==1.4.3
dj-database-url==0.4.0
Django==1.8.13
django-appconf==1.0.2
django-crispy-forms==1.6.0
django-el-pagination==2.1.2
django-imagekit==3.3
django-location-field==2.0.0b0
django-registration==2.1
django-storages==1.4.1
gunicorn==19.4.5
html5lib==0.9999999
pilkit==1.1.13
Pillow==3.2.0
psycopg2==2.6.1
python-decouple==3.0
six==1.10.0
Unipath==1.1
whitenoise==2.0.6
